Create new zim using the new namespace scheme.
All items added by the user are put in the `C` namespace (for "Content").

Path given to the creator do not need to include the namespace as the
user can't specify a namespace.
However, path containing a namespace will not break. The namespace will
simply be handled as part of the path, as a subdirectory.

MinorVersion us bumped to 1.